+++ Examples/EmbraceIOTestApp/EmbraceIOTestApp/Common/Protocols/TestScreenDataModel.swift
@@ -1,0 +1,14 @@
+//
+//  TestScreenDataModel.swift
+//  EmbraceIOTestApp
+//
+//
+
+import SwiftUI
+
+protocol TestScreenDataModel {
+    associatedtype UIComponent: View
+    var title: String { get }
+    var identifier: String { get }
+    @ViewBuilder var uiComponent: UIComponent { get }
+}