+++ Sources/EmbraceCore/Internal/Tracing/StorageSpanExporter.swift
@@ -11,12 +11,14 @@ import OpenTelemetrySdk
 class StorageSpanExporter: SpanExporter {
 
     private(set) weak var storage: EmbraceStorage?
+    private(set) weak var sessionController: SessionControllable?
     private weak var logger: InternalLogger?
 
     let validation: SpanDataValidation
 
     init(options: Options, logger: InternalLogger) {
         self.storage = options.storage
+        self.sessionController = options.sessionController
         self.validation = SpanDataValidation(validators: options.validators)
         self.logger = logger
     }
@@ -46,7 +48,6 @@ class StorageSpanExporter: SpanExporter {
     }
 
     public func flush(explicitTimeout: TimeInterval?) -> SpanExporterResultCode {
-        // TODO: do we need to make sure storage writes are finished?
         return .success
     }
 
@@ -72,6 +73,8 @@ extension StorageSpanExporter {
             type: spanData.embType,
             data: data,
             startTime: spanData.startTime,
-            endTime: endTime )
+            endTime: endTime,
+            sessionIdentifier: sessionController?.currentSession?.id
+        )
     }
 }