+++ Sources/EmbraceCore/Options/Embrace+Endpoints.swift
@@ -13,11 +13,21 @@ extension Embrace {
         /// The base URL to upload session data
         @objc public let baseURL: String
 
-        /// The base URL to upload session data while a debugger is attached
-        @objc public let developmentBaseURL: String
-
         /// The base URL to retrieve remote config
         @objc public let configBaseURL: String
+
+        @available(*, deprecated, message: "Development base URL is not used anymore.")
+        @objc public let developmentBaseURL: String = "" // TODO: Remove in next major version
+
+        /// Initializer that allows for custom endpoints.
+        /// - Note: If you wish to use the default endpoints please refer to the convenience initializer: `init(appId: String)`.
+        /// - Parameters:
+        ///   - baseURL: Endpoint for session data upload
+        ///   - configBaseURL: Endpoint to fetch the remote config
+        @objc public init(baseURL: String, configBaseURL: String) {
+            self.baseURL = baseURL
+            self.configBaseURL = configBaseURL
+        }
 
         /// Initializer that allows for custom endpoints.
         /// - Note: If you wish to use the default endpoints please refer to the convenience initializer: `init(appId: String)`.
@@ -25,9 +35,10 @@ extension Embrace {
         ///   - baseURL: Endpoint for session data upload
         ///   - developmentBaseURL: Endpoint for session data upload while debugging
         ///   - configBaseURL: Endpoint to fetch the remote config
+        @available(*, deprecated, message: "Use `init(baseURL:configBaseURL)` instead.")
         @objc public init(baseURL: String, developmentBaseURL: String, configBaseURL: String) {
+            // TODO: Remove in next major version
             self.baseURL = baseURL
-            self.developmentBaseURL = developmentBaseURL
             self.configBaseURL = configBaseURL
         }
     }
@@ -39,7 +50,7 @@ extension Embrace.Endpoints {
     @objc convenience init(appId: String) {
         self.init(
             baseURL: "https://a-\(appId).data.emb-api.com",
-            developmentBaseURL: "https://data-dev.emb-api.com",
-            configBaseURL: "https://a-\(appId).config.emb-api.com")
+            configBaseURL: "https://a-\(appId).config.emb-api.com"
+        )
     }
 }