+++ Sources/EmbraceCore/Public/Embrace+OTel.swift
@@ -12,7 +12,7 @@ import OpenTelemetrySdk
 extension Embrace: EmbraceOpenTelemetry {
     private var exporter: SpanExporter {
         StorageSpanExporter(
-            options: .init(storage: storage),
+            options: .init(storage: storage, sessionController: sessionController),
             logger: Embrace.logger
         )
     }
@@ -119,6 +119,8 @@ extension Embrace: EmbraceOpenTelemetry {
     ///   - severity: `LogSeverity` for the log.
     ///   - attributes: Attributes for the log.
     ///   - stackTraceBehavior: Defines if the stack trace information should be added to the log
+    ///
+    /// - Important: Only `warn` and `error` logs will have stacktraces.
     public func log(
         _ message: String,
         severity: LogSeverity,
@@ -143,6 +145,8 @@ extension Embrace: EmbraceOpenTelemetry {
     ///   - timestamp: Timestamp for the log.
     ///   - attributes: Attributes for the log.
     ///   - stackTraceBehavior: Defines if the stack trace information should be added to the log
+    ///
+    /// - Important: Only `warn` and `error` logs will have stacktraces.
     public func log(
         _ message: String,
         severity: LogSeverity,
@@ -173,6 +177,8 @@ extension Embrace: EmbraceOpenTelemetry {
     ///   - attachment: Data of the attachment
     ///   - attributes: Attributes for the log.
     ///   - stackTraceBehavior: Defines if the stack trace information should be added to the log
+    ///
+    /// - Important: Only `warn` and `error` logs will have stacktraces.
     public func log(
         _ message: String,
         severity: LogSeverity,
@@ -205,6 +211,8 @@ extension Embrace: EmbraceOpenTelemetry {
     ///   - attachmentUrl: URL to dowload the attachment data
     ///   - attributes: Attributes for the log.
     ///   - stackTraceBehavior: Defines if the stack trace information should be added to the log
+    ///
+    /// - Important: Only `warn` and `error` logs will have stacktraces.
     public func log(
         _ message: String,
         severity: LogSeverity,