+++ Sources/EmbraceStorageInternal/EmbraceStorageError.swift
@@ -7,7 +7,6 @@ import GRDB
 
 public enum EmbraceStorageError: Error, Equatable {
     case cannotUpsertSpan(spanName: String, message: String)
-    // TODO: Add missing errors in here
 }
 
 extension EmbraceStorageError: LocalizedError, CustomNSError {