+++ Tests/EmbraceCoreTests/Capture/Network/Proxy/URLSessionDelegateProxyAsTaskDelegateTests.swift
@@ -91,7 +91,7 @@ final class URLSessionDelegateProxyAsTaskDelegateTests: XCTestCase {
         try otherSwizzler?.install()
     }
 
-    @available(iOS 15.0, watchOS 8.0, *)
+    @available(iOS 15.0, watchOS 8.0, tvOS 15.0, *)
     func testWithPreviousSwizzledProxy_taskWithDelegate_callsTaskDelegateOnly() throws {
         /*
          This test checks what happens when there's a delegate assigned when creating a
@@ -129,7 +129,7 @@ final class URLSessionDelegateProxyAsTaskDelegateTests: XCTestCase {
         unswizzleOtherSwizzler()
     }
 
-    @available(iOS 15.0, watchOS 8.0, *)
+    @available(iOS 15.0, watchOS 8.0, tvOS 15.0, *)
     func testWithProxySwizzledAfterEmbrace_taskWithDelegate_callsTaskDelegateOnly() throws {
         /*
          This test checks what happens when there's a delegate assigned when creating a
@@ -167,7 +167,7 @@ final class URLSessionDelegateProxyAsTaskDelegateTests: XCTestCase {
         unswizzleDefaultCaptureService()
     }
 
-    @available(iOS 15.0, watchOS 8.0, *)
+    @available(iOS 15.0, watchOS 8.0, tvOS 15.0, *)
     func test_taskWithDelegate_callsTaskDelegateOnly() throws {
         givenCaptureServiceInstalled()
         givenTaskDelegate()
@@ -192,7 +192,7 @@ final class URLSessionDelegateProxyAsTaskDelegateTests: XCTestCase {
         unswizzleDefaultCaptureService()
     }
 
-    @available(iOS 15.0, watchOS 8.0, *)
+    @available(iOS 15.0, watchOS 8.0, tvOS 15.0, *)
     func test_taskWithDelegate_thatDoesNotImplementMethods_callsSessionDelegateOnly() throws {
         givenCaptureServiceInstalled()
         givenTaskDelegate(NotImplementedURLSessionDelegate())
@@ -213,7 +213,7 @@ final class URLSessionDelegateProxyAsTaskDelegateTests: XCTestCase {
         unswizzleDefaultCaptureService()
     }
 
-    @available(iOS 15.0, watchOS 8.0, *)
+    @available(iOS 15.0, watchOS 8.0, tvOS 15.0, *)
     func test_async_taskWithNoDelegate_callsSessionDelegate() async throws {
         givenCaptureServiceInstalled()
         givenSessionDelegate()
@@ -232,7 +232,7 @@ final class URLSessionDelegateProxyAsTaskDelegateTests: XCTestCase {
         unswizzleDefaultCaptureService()
     }
 
-    @available(iOS 15.0, watchOS 8.0, *)
+    @available(iOS 15.0, watchOS 8.0, tvOS 15.0, *)
     func test_async_taskWithDelegate_doesNotCallSessionDelegate() async throws {
         givenCaptureServiceInstalled()